Fix compile error - signed vs. unsigned comparison.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@10479 7a7537e8-13f0-0310-91df-b6672ffda945

+++ scheduler/cups-driverd.cxx
@@ -671,7 +671,7 @@ cat_tar(const char *name,		/* I - PPD name */
 
       for (total = 0; total < curinfo.st_size; total += bytes)
       {
-        if ((bytes = (curinfo.st_size - total)) > sizeof(buffer))
+        if ((size_t)(bytes = (curinfo.st_size - total)) > sizeof(buffer))
           bytes = sizeof(buffer);
 
         if ((bytes = cupsFileRead(fp, buffer, bytes)) < 0)
